+++ com.microej.example.iot.ssl.mqtt.subscriber/src/main/java/com/microej/example/iot/mqtt/app/SubscriberBG.java
@@ -1,12 +1,14 @@
 /*
  * Java
  *
- * Copyright 2016 IS2T. All rights reserved.
- * Use of this source code is subject to license terms.
+ * Copyright 2016-2018 IS2T. All rights reserved.
+ * For demonstration purpose only.
+ * IS2T PROPRIETARY. Use is subject to license terms.
  */
 package com.microej.example.iot.mqtt.app;
 
 import com.microej.example.iot.mqtt.HelloWorldSubscriber;
+
 import ej.wadapps.app.BackgroundService;
 
 /**
@@ -14,15 +16,18 @@ import ej.wadapps.app.BackgroundService;
  */
 public class SubscriberBG implements BackgroundService {
 
+	private HelloWorldSubscriber helloWorldSubscriber;
+
 	@Override
 	public void onStart() {
-		HelloWorldSubscriber.main(new String[0]);
-
+		helloWorldSubscriber = new HelloWorldSubscriber();
+		helloWorldSubscriber.start();
 	}
 
 	@Override
 	public void onStop() {
-
+		helloWorldSubscriber.stop();
+		helloWorldSubscriber = null;
 	}
 
 }