+++ com.microej.example.iot.ssl.mqtt.subscriber/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 2.1.0 (2018-08-01)
+
+Bug Fixes:
+  - Fix unregistering on lost
+  
+Features:
+  - Use connectivity manager 1.1.0.
+  - Add default connectivity manager for standalone usage
+  
 ## 2.0.0 (2016-12-29)
 Features:
   - Change organization.
@@ -10,3 +19,9 @@ Features:
 Features:
   - Initial revision.
 
+<!--
+    Markdown
+    Copyright 2016-2018 IS2T. All rights reserved.
+    For demonstration purpose only.
+    IS2T PROPRIETARY. Use is subject to license terms.
+-->