+++ com.microej.example.iot.ssl.rest/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.1.0 (2018-08-01)
+
+Features:
+  - Update to latest Widget.
+  - Use connectivity manager 1.1.0.
+  - Add default connectivity manager for standalone usage
+  
 ## 2.0.1 (2017-08-07)
 Bug Fixes:
   -  Update the web site for testing (post-echoman.org)
@@ -10,3 +17,9 @@ Features:
 Features:
   - Initial revision.
 
+<!--
+    Markdown
+    Copyright 2016-2018 IS2T. All rights reserved.
+    For demonstration purpose only.
+    IS2T PROPRIETARY. Use is subject to license terms.
+-->